gem5-resources: x86 ubuntu 24.04 disk image (#34)
Updating scripts to build ubuntu 24.04.

+++ src/x86-ubuntu/scripts/post-installation.sh
@@ -69,7 +69,11 @@ echo "Done building and installing gem5-bridge (m5) and libm5"
 # Disable network by default
 echo "Disabling network by default"
 echo "See README.md for instructions on how to enable network"
-mv /etc/netplan/00-installer-config.yaml /etc/netplan/00-installer-config.yaml.bak
-netplan apply
+if [ -f /etc/netplan/50-cloud-init.yaml ]; then
+    mv /etc/netplan/50-cloud-init.yaml /etc/netplan/50-cloud-init.yaml.bak
+elif [ -f /etc/netplan/00-installer-config.yaml ]; then
+    mv /etc/netplan/00-installer-config.yaml /etc/netplan/00-installer-config.yaml.bak
+    netplan apply
+fi
 
 echo "Post Installation Done"
